backend/requirements.txt
opencv-python
easyocr
numpy
flask
python-docx
xlsxwriter
